updated logger and handled blank and empty responses
