+++ src/test/java/io/github/mfaisalkhatri/api/restfulecommerce/BaseTest.java
@@ -7,6 +7,7 @@ import com.microsoft.playwright.APIRequest;
 import com.microsoft.playwright.APIRequestContext;
 import com.microsoft.playwright.APIResponse;
 import com.microsoft.playwright.Playwright;
+import io.github.mfaisalkhatri.api.logger.Logger;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
